refactor(globals): split register modules into granular subpaths
Split monolithic register files into per-feature subpaths so that
--globals auto injects only the code for identifiers actually used.

Packages split:
- dom-elements: 8 subpaths (document, canvas, image, observers,
  font-face, match-media, location, navigator)
- node-globals: 7 subpaths (process, buffer, timers, encoding,
  url, structured-clone, microtask)
- dom-events: 3 subpaths (event-target, custom-events, ui-events)
- streams: 5 subpaths (readable, writable, transform, text-streams,
  queuing)
- fetch: 1 subpath (fetch — all 4 identifiers tightly coupled)
- web-globals: 2 subpaths (performance, formdata — no longer pull
  the entire web API surface)

Bundle size improvements (auto mode):
- excalibur-jelly-jumper: 2.95 MB → 2.61 MB (-12%)
- net-express-hello: 2.02 MB → 1.86 MB (-8%)

Catch-all ./register paths remain for backwards compatibility.
