+++ packages/dom/dom-elements/src/register/image.ts
@@ -1,0 +1,8 @@
+// Registers: Image, HTMLImageElement
+
+import { HTMLImageElement } from '../html-image-element.js';
+import { Image } from '../image.js';
+import { defineGlobal } from './helpers.js';
+
+defineGlobal('HTMLImageElement', HTMLImageElement);
+defineGlobal('Image', Image);
