+++ packages/dom/dom-elements/src/register/match-media.ts
@@ -1,0 +1,6 @@
+// Registers: matchMedia
+
+import { matchMedia } from '../match-media.js';
+import { defineGlobalIfMissing } from './helpers.js';
+
+defineGlobalIfMissing('matchMedia', matchMedia);
