+++ packages/infra/resolve-npm/lib/globals-map.mjs
@@ -5,11 +5,15 @@
  * to resolve the user's explicit list of globals into the set of `/register`
  * modules that must be injected into the bundle.
  *
- * The register paths use bare specifiers ("fetch/register" rather than
- * "@gjsify/fetch/register") so they go through the normal ALIASES_WEB_FOR_GJS
+ * The register paths use bare specifiers ("fetch/register/fetch" rather than
+ * "@gjsify/fetch/register/fetch") so they go through the normal ALIASES_WEB_FOR_GJS
  * / ALIASES_WEB_FOR_NODE resolution — which means they become no-ops on Node
  * (where native globals already exist) and resolve to the real
- * `@gjsify/<pkg>/register` modules on GJS.
+ * `@gjsify/<pkg>/register/<subpath>` modules on GJS.
+ *
+ * Granular subpaths (e.g. "dom-events/register/event-target" vs "dom-events/register")
+ * ensure that --globals auto injects only the register code for identifiers
+ * actually referenced in the bundle, keeping unused globals out of the output.
  */
 /**
  * Pre-defined groups of identifiers that can be used as shorthand in `--globals`.
@@ -56,100 +60,100 @@ export const GJS_GLOBALS_GROUPS = {
 };
 
 export const GJS_GLOBALS_MAP = {
-    // --- Node.js globals ------------------------------------------------
-    Buffer:               '@gjsify/buffer/register',
-    process:              '@gjsify/node-globals/register',
-    setImmediate:         '@gjsify/node-globals/register',
-    clearImmediate:       '@gjsify/node-globals/register',
-    queueMicrotask:       '@gjsify/node-globals/register',
-    structuredClone:      '@gjsify/node-globals/register',
-    btoa:                 '@gjsify/node-globals/register',
-    atob:                 '@gjsify/node-globals/register',
+    // --- Node.js globals (granular register subpaths) ----------------------
+    Buffer:               '@gjsify/node-globals/register/buffer',
+    process:              '@gjsify/node-globals/register/process',
+    setImmediate:         '@gjsify/node-globals/register/timers',
+    clearImmediate:       '@gjsify/node-globals/register/timers',
+    queueMicrotask:       '@gjsify/node-globals/register/microtask',
+    structuredClone:      '@gjsify/node-globals/register/structured-clone',
+    btoa:                 '@gjsify/node-globals/register/encoding',
+    atob:                 '@gjsify/node-globals/register/encoding',
 
-    // --- URL (shared Node + Web) ---------------------------------------
-    URL:                  '@gjsify/node-globals/register',
-    URLSearchParams:      '@gjsify/node-globals/register',
+    // --- URL (shared Node + Web) -------------------------------------------
+    URL:                  '@gjsify/node-globals/register/url',
+    URLSearchParams:      '@gjsify/node-globals/register/url',
 
-    // --- Blob / File (exposed via node:buffer and Web File API) --------
+    // --- Blob / File (exposed via node:buffer and Web File API) ------------
     Blob:                 '@gjsify/buffer/register',
     File:                 '@gjsify/buffer/register',
 
-    // --- Fetch stack ----------------------------------------------------
-    fetch:                'fetch/register',
-    Headers:              'fetch/register',
-    Request:              'fetch/register',
-    Response:             'fetch/register',
+    // --- Fetch stack -------------------------------------------------------
+    fetch:                'fetch/register/fetch',
+    Headers:              'fetch/register/fetch',
+    Request:              'fetch/register/fetch',
+    Response:             'fetch/register/fetch',
 
-    // --- FormData ------------------------------------------------------
-    FormData:             '@gjsify/web-globals/register',
+    // --- FormData ----------------------------------------------------------
+    FormData:             '@gjsify/web-globals/register/formdata',
 
-    // --- WHATWG Streams ------------------------------------------------
-    ReadableStream:       'web-streams/register',
-    WritableStream:       'web-streams/register',
-    TransformStream:      'web-streams/register',
-    TextEncoderStream:    'web-streams/register',
-    TextDecoderStream:    'web-streams/register',
-    ByteLengthQueuingStrategy: 'web-streams/register',
-    CountQueuingStrategy: 'web-streams/register',
+    // --- WHATWG Streams (granular register subpaths) ------------------------
+    ReadableStream:       'web-streams/register/readable',
+    WritableStream:       'web-streams/register/writable',
+    TransformStream:      'web-streams/register/transform',
+    TextEncoderStream:    'web-streams/register/text-streams',
+    TextDecoderStream:    'web-streams/register/text-streams',
+    ByteLengthQueuingStrategy: 'web-streams/register/queuing',
+    CountQueuingStrategy: 'web-streams/register/queuing',
 
-    // --- Compression Streams -------------------------------------------
+    // --- Compression Streams -----------------------------------------------
     CompressionStream:    'compression-streams/register',
     DecompressionStream:  'compression-streams/register',
 
-    // --- WebCrypto -----------------------------------------------------
+    // --- WebCrypto ---------------------------------------------------------
     crypto:               'webcrypto/register',
 
-    // --- Events + Abort ------------------------------------------------
+    // --- Events + Abort (granular register subpaths) -----------------------
     AbortController:      'abort-controller/register',
     AbortSignal:          'abort-controller/register',
-    Event:                'dom-events/register',
-    EventTarget:          'dom-events/register',
-    CustomEvent:          'dom-events/register',
-    MessageEvent:         'dom-events/register',
-    ErrorEvent:           'dom-events/register',
-    CloseEvent:           'dom-events/register',
-    ProgressEvent:        'dom-events/register',
-    UIEvent:              'dom-events/register',
-    MouseEvent:           'dom-events/register',
-    PointerEvent:         'dom-events/register',
-    KeyboardEvent:        'dom-events/register',
-    WheelEvent:           'dom-events/register',
-    FocusEvent:           'dom-events/register',
+    Event:                'dom-events/register/event-target',
+    EventTarget:          'dom-events/register/event-target',
+    CustomEvent:          'dom-events/register/custom-events',
+    MessageEvent:         'dom-events/register/custom-events',
+    ErrorEvent:           'dom-events/register/custom-events',
+    CloseEvent:           'dom-events/register/custom-events',
+    ProgressEvent:        'dom-events/register/custom-events',
+    UIEvent:              'dom-events/register/ui-events',
+    MouseEvent:           'dom-events/register/ui-events',
+    PointerEvent:         'dom-events/register/ui-events',
+    KeyboardEvent:        'dom-events/register/ui-events',
+    WheelEvent:           'dom-events/register/ui-events',
+    FocusEvent:           'dom-events/register/ui-events',
     EventSource:          'eventsource/register',
 
-    // --- Performance ---------------------------------------------------
-    performance:          '@gjsify/web-globals/register',
-    PerformanceObserver:  '@gjsify/web-globals/register',
+    // --- Performance -------------------------------------------------------
+    performance:          '@gjsify/web-globals/register/performance',
+    PerformanceObserver:  '@gjsify/web-globals/register/performance',
 
-    // --- DOMException --------------------------------------------------
+    // --- DOMException ------------------------------------------------------
     DOMException:         'dom-exception/register',
 
-    // --- XMLHttpRequest + URL.createObjectURL --------------------------
+    // --- XMLHttpRequest + URL.createObjectURL ------------------------------
     XMLHttpRequest:       '@gjsify/xmlhttprequest/register',
 
-    // --- DOMParser -----------------------------------------------------
+    // --- DOMParser ---------------------------------------------------------
     DOMParser:            '@gjsify/domparser/register',
 
-    // --- Web Audio API (GStreamer backend) --------------------------------
+    // --- Web Audio API (GStreamer backend) ----------------------------------
     AudioContext:         '@gjsify/webaudio/register',
     webkitAudioContext:   '@gjsify/webaudio/register',
     Audio:                '@gjsify/webaudio/register',
     HTMLAudioElement:     '@gjsify/webaudio/register',
 
-    // --- Gamepad API (libmanette backend) --------------------------------
+    // --- Gamepad API (libmanette backend) -----------------------------------
     GamepadEvent:         '@gjsify/gamepad/register',
 
-    // --- DOM elements (browser-compat) ---------------------------------
-    document:             '@gjsify/dom-elements/register',
-    Image:                '@gjsify/dom-elements/register',
-    HTMLCanvasElement:    '@gjsify/dom-elements/register',
-    HTMLImageElement:     '@gjsify/dom-elements/register',
-    HTMLElement:          '@gjsify/dom-elements/register',
-    MutationObserver:     '@gjsify/dom-elements/register',
-    ResizeObserver:       '@gjsify/dom-elements/register',
-    IntersectionObserver: '@gjsify/dom-elements/register',
-    FontFace:             '@gjsify/dom-elements/register',
-    matchMedia:           '@gjsify/dom-elements/register',
-    location:             '@gjsify/dom-elements/register',
-    navigator:            '@gjsify/dom-elements/register',
+    // --- DOM elements (granular register subpaths) -------------------------
+    document:             '@gjsify/dom-elements/register/document',
+    Image:                '@gjsify/dom-elements/register/image',
+    HTMLCanvasElement:    '@gjsify/dom-elements/register/canvas',
+    HTMLImageElement:     '@gjsify/dom-elements/register/image',
+    HTMLElement:          '@gjsify/dom-elements/register/document',
+    MutationObserver:     '@gjsify/dom-elements/register/observers',
+    ResizeObserver:       '@gjsify/dom-elements/register/observers',
+    IntersectionObserver: '@gjsify/dom-elements/register/observers',
+    FontFace:             '@gjsify/dom-elements/register/font-face',
+    matchMedia:           '@gjsify/dom-elements/register/match-media',
+    location:             '@gjsify/dom-elements/register/location',
+    navigator:            '@gjsify/dom-elements/register/navigator',
 };
