+++ packages/infra/resolve-npm/lib/index.mjs
@@ -155,11 +155,20 @@ export const ALIASES_WEB_FOR_GJS = {
     'abort-controller/register': '@gjsify/abort-controller/register',
     'compression-streams/register': '@gjsify/compression-streams/register',
     'dom-events/register': '@gjsify/dom-events/register',
+    'dom-events/register/event-target': '@gjsify/dom-events/register/event-target',
+    'dom-events/register/custom-events': '@gjsify/dom-events/register/custom-events',
+    'dom-events/register/ui-events': '@gjsify/dom-events/register/ui-events',
     'dom-exception/register': '@gjsify/dom-exception/register',
     'eventsource/register': '@gjsify/eventsource/register',
     'fetch/register': '@gjsify/fetch/register',
+    'fetch/register/fetch': '@gjsify/fetch/register/fetch',
     'webcrypto/register': '@gjsify/webcrypto/register',
     'web-streams/register': '@gjsify/web-streams/register',
+    'web-streams/register/readable': '@gjsify/web-streams/register/readable',
+    'web-streams/register/writable': '@gjsify/web-streams/register/writable',
+    'web-streams/register/transform': '@gjsify/web-streams/register/transform',
+    'web-streams/register/text-streams': '@gjsify/web-streams/register/text-streams',
+    'web-streams/register/queuing': '@gjsify/web-streams/register/queuing',
 
     // xmlhttprequest + DOMParser
     'xmlhttprequest': '@gjsify/xmlhttprequest',
@@ -180,8 +189,17 @@ export const ALIASES_WEB_FOR_GJS = {
 export const ALIASES_GENERAL_FOR_NODE = {
     '@gjsify/node-globals': '@gjsify/empty',
     '@gjsify/node-globals/register': '@gjsify/empty',
+    '@gjsify/node-globals/register/process': '@gjsify/empty',
+    '@gjsify/node-globals/register/buffer': '@gjsify/empty',
+    '@gjsify/node-globals/register/timers': '@gjsify/empty',
+    '@gjsify/node-globals/register/encoding': '@gjsify/empty',
+    '@gjsify/node-globals/register/url': '@gjsify/empty',
+    '@gjsify/node-globals/register/structured-clone': '@gjsify/empty',
+    '@gjsify/node-globals/register/microtask': '@gjsify/empty',
     '@gjsify/web-globals': '@gjsify/empty',
     '@gjsify/web-globals/register': '@gjsify/empty',
+    '@gjsify/web-globals/register/performance': '@gjsify/empty',
+    '@gjsify/web-globals/register/formdata': '@gjsify/empty',
 }
 
 /** Record of Gjs modules (build in or not) and his replacement for Node */
@@ -209,20 +227,46 @@ export const ALIASES_WEB_FOR_NODE = {
     'abort-controller/register': '@gjsify/empty',
     'compression-streams/register': '@gjsify/empty',
     'dom-events/register': '@gjsify/empty',
+    'dom-events/register/event-target': '@gjsify/empty',
+    'dom-events/register/custom-events': '@gjsify/empty',
+    'dom-events/register/ui-events': '@gjsify/empty',
     'dom-exception/register': '@gjsify/empty',
     'eventsource/register': '@gjsify/empty',
     'fetch/register': '@gjsify/empty',
+    'fetch/register/fetch': '@gjsify/empty',
     'webcrypto/register': '@gjsify/empty',
     'web-streams/register': '@gjsify/empty',
+    'web-streams/register/readable': '@gjsify/empty',
+    'web-streams/register/writable': '@gjsify/empty',
+    'web-streams/register/transform': '@gjsify/empty',
+    'web-streams/register/text-streams': '@gjsify/empty',
+    'web-streams/register/queuing': '@gjsify/empty',
     '@gjsify/abort-controller/register': '@gjsify/empty',
     '@gjsify/compression-streams/register': '@gjsify/empty',
     '@gjsify/dom-events/register': '@gjsify/empty',
+    '@gjsify/dom-events/register/event-target': '@gjsify/empty',
+    '@gjsify/dom-events/register/custom-events': '@gjsify/empty',
+    '@gjsify/dom-events/register/ui-events': '@gjsify/empty',
     '@gjsify/dom-exception/register': '@gjsify/empty',
     '@gjsify/eventsource/register': '@gjsify/empty',
     '@gjsify/fetch/register': '@gjsify/empty',
+    '@gjsify/fetch/register/fetch': '@gjsify/empty',
     '@gjsify/webcrypto/register': '@gjsify/empty',
     '@gjsify/web-streams/register': '@gjsify/empty',
+    '@gjsify/web-streams/register/readable': '@gjsify/empty',
+    '@gjsify/web-streams/register/writable': '@gjsify/empty',
+    '@gjsify/web-streams/register/transform': '@gjsify/empty',
+    '@gjsify/web-streams/register/text-streams': '@gjsify/empty',
+    '@gjsify/web-streams/register/queuing': '@gjsify/empty',
     '@gjsify/dom-elements/register': '@gjsify/empty',
+    '@gjsify/dom-elements/register/document': '@gjsify/empty',
+    '@gjsify/dom-elements/register/canvas': '@gjsify/empty',
+    '@gjsify/dom-elements/register/image': '@gjsify/empty',
+    '@gjsify/dom-elements/register/observers': '@gjsify/empty',
+    '@gjsify/dom-elements/register/font-face': '@gjsify/empty',
+    '@gjsify/dom-elements/register/match-media': '@gjsify/empty',
+    '@gjsify/dom-elements/register/location': '@gjsify/empty',
+    '@gjsify/dom-elements/register/navigator': '@gjsify/empty',
     '@gjsify/buffer/register': '@gjsify/empty',
 
     // xmlhttprequest + DOMParser — no-op on Node
