+++ packages/node/globals/package.json
@@ -17,12 +17,33 @@
             "default": "./lib/esm/index.js"
         },
         "./register": {
-            "types": "./lib/types/register.d.ts",
             "default": "./lib/esm/register.js"
+        },
+        "./register/process": {
+            "default": "./lib/esm/register/process.js"
+        },
+        "./register/buffer": {
+            "default": "./lib/esm/register/buffer.js"
+        },
+        "./register/timers": {
+            "default": "./lib/esm/register/timers.js"
+        },
+        "./register/encoding": {
+            "default": "./lib/esm/register/encoding.js"
+        },
+        "./register/url": {
+            "default": "./lib/esm/register/url.js"
+        },
+        "./register/structured-clone": {
+            "default": "./lib/esm/register/structured-clone.js"
+        },
+        "./register/microtask": {
+            "default": "./lib/esm/register/microtask.js"
         }
     },
     "sideEffects": [
-        "./lib/esm/register.js"
+        "./lib/esm/register.js",
+        "./lib/esm/register/*.js"
     ],
     "scripts": {
         "clear": "rm -rf lib tsconfig.tsbuildinfo tsconfig.types.tsbuildinfo test.gjs.mjs test.node.mjs || exit 0",
