+++ packages/node/globals/src/register/structured-clone.ts
@@ -1,0 +1,12 @@
+// Registers: structuredClone
+
+import { structuredClone as structuredClonePolyfill } from '@gjsify/utils';
+
+if (typeof globalThis.structuredClone !== 'function') {
+  Object.defineProperty(globalThis, 'structuredClone', {
+    value: structuredClonePolyfill,
+    enumerable: true,
+    writable: true,
+    configurable: true,
+  });
+}
