+++ packages/node/globals/src/register/timers.ts
@@ -1,0 +1,27 @@
+// Registers: setImmediate, clearImmediate
+
+function setImmediate<T extends any[]>(callback: (...args: T) => void, ...args: T): ReturnType<typeof setTimeout> {
+  return setTimeout(callback, 0, ...args);
+}
+
+function clearImmediate(id: ReturnType<typeof setTimeout>): void {
+  clearTimeout(id);
+}
+
+if (!('setImmediate' in globalThis)) {
+  Object.defineProperty(globalThis, 'setImmediate', {
+    value: setImmediate,
+    enumerable: true,
+    writable: true,
+    configurable: true,
+  });
+}
+
+if (!('clearImmediate' in globalThis)) {
+  Object.defineProperty(globalThis, 'clearImmediate', {
+    value: clearImmediate,
+    enumerable: true,
+    writable: true,
+    configurable: true,
+  });
+}
