+++ packages/web/dom-events/package.json
@@ -11,13 +11,22 @@
             "default": "./lib/esm/index.js"
         },
         "./register": {
-            "types": "./lib/types/register.d.ts",
             "default": "./lib/esm/register.js"
+        },
+        "./register/event-target": {
+            "default": "./lib/esm/register/event-target.js"
+        },
+        "./register/custom-events": {
+            "default": "./lib/esm/register/custom-events.js"
+        },
+        "./register/ui-events": {
+            "default": "./lib/esm/register/ui-events.js"
         },
         "./globals": "./globals.mjs"
     },
     "sideEffects": [
         "./lib/esm/register.js",
+        "./lib/esm/register/*.js",
         "./globals.mjs"
     ],
     "scripts": {
