+++ packages/web/fetch/package.json
@@ -11,13 +11,16 @@
             "default": "./lib/esm/index.js"
         },
         "./register": {
-            "types": "./lib/types/register.d.ts",
             "default": "./lib/esm/register.js"
+        },
+        "./register/fetch": {
+            "default": "./lib/esm/register/fetch.js"
         },
         "./globals": "./globals.mjs"
     },
     "sideEffects": [
         "./lib/esm/register.js",
+        "./lib/esm/register/*.js",
         "./globals.mjs"
     ],
     "scripts": {
