+++ packages/web/streams/src/register/writable.ts
@@ -1,0 +1,16 @@
+// Registers: WritableStream
+
+import { WritableStream } from '../index.js';
+
+function isNativeStreamUsable(Ctor: unknown, method: string): boolean {
+  try {
+    if (typeof Ctor !== 'function') return false;
+    return typeof (Ctor as any).prototype[method] === 'function';
+  } catch {
+    return false;
+  }
+}
+
+if (!isNativeStreamUsable(globalThis.WritableStream, 'getWriter')) {
+  (globalThis as any).WritableStream = WritableStream;
+}
