+++ packages/web/web-globals/src/register/formdata.ts
@@ -1,0 +1,7 @@
+// Registers: FormData
+
+import { FormData } from '@gjsify/formdata';
+
+if (typeof globalThis.FormData !== 'function') {
+  (globalThis as any).FormData = FormData;
+}
